requirements.txt
pycryptodome>=3.19.0
psutil>=5.9.0 